+++ jpo-ode-plugins/src/test/java/us/dot/its/jpo/ode/plugin/j2735/J2735IntersectionAccessPointTest.java
@@ -1,8 +1,8 @@
 package us.dot.its.jpo.ode.plugin.j2735;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.Assert.assertEquals;
 
-import org.junit.jupiter.api.Test;
+import org.junit.Test;
 
 public class J2735IntersectionAccessPointTest {
     
@@ -10,12 +10,12 @@ public class J2735IntersectionAccessPointTest {
 	public void testGettersSetters() {
         J2735IntersectionAccessPoint interAccessPoint = new J2735IntersectionAccessPoint();
         interAccessPoint.setApproach(5);
-        assertEquals(interAccessPoint.getApproach(), 5);
+        assertEquals(interAccessPoint.getApproach().intValue(), 5);
 
         interAccessPoint.setConnection(4);
-        assertEquals(interAccessPoint.getConnection(), 4);
+        assertEquals(interAccessPoint.getConnection().intValue(), 4);
 
         interAccessPoint.setLane(3);
-        assertEquals(interAccessPoint.getLane(), 3);
+        assertEquals(interAccessPoint.getLane().intValue(), 3);
 	}
 }
