+++ jpo-ode-plugins/src/test/java/us/dot/its/jpo/ode/plugin/j2735/J2735IntersectionReferenceIDTest.java
@@ -1,8 +1,8 @@
 package us.dot.its.jpo.ode.plugin.j2735;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.Assert.assertEquals;
 
-import org.junit.jupiter.api.Test;
+import org.junit.Test;
 
 public class J2735IntersectionReferenceIDTest {
     
@@ -10,9 +10,9 @@ public class J2735IntersectionReferenceIDTest {
 	public void testGettersSetters() {
         J2735IntersectionReferenceID interRefId = new J2735IntersectionReferenceID();
         interRefId.setId(5);
-        assertEquals(interRefId.getId(), 5);
+        assertEquals(interRefId.getId().intValue(), 5);
 
         interRefId.setRegion(4);
-        assertEquals(interRefId.getRegion(), 4);
+        assertEquals(interRefId.getRegion().intValue(), 4);
 	}
 }
