+++ jpo-ode-plugins/src/test/java/us/dot/its/jpo/ode/plugin/j2735/J2735SignalRequestTest.java
@@ -1,8 +1,8 @@
 package us.dot.its.jpo.ode.plugin.j2735;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.Assert.assertEquals;
 
-import org.junit.jupiter.api.Test;
+import org.junit.Test;
 
 public class J2735SignalRequestTest {
     
@@ -13,10 +13,10 @@ public class J2735SignalRequestTest {
         J2735IntersectionReferenceID interRefId = new J2735IntersectionReferenceID();
         interRefId.setId(5);
         signalRequest.setId(interRefId);
-        assertEquals(signalRequest.getId().getId(), 5);
+        assertEquals(signalRequest.getId().getId().intValue(), 5);
 
         signalRequest.setRequestID(5);
-        assertEquals(signalRequest.getRequestID(), 5);
+        assertEquals(signalRequest.getRequestID().intValue(), 5);
 
         signalRequest.setRequestType(J2735PriorityRequestType.priorityRequest);
         assertEquals(signalRequest.getRequestType(), J2735PriorityRequestType.priorityRequest);
@@ -24,9 +24,9 @@ public class J2735SignalRequestTest {
         J2735IntersectionAccessPoint interAccessPoint = new J2735IntersectionAccessPoint();
         interAccessPoint.setApproach(5);
         signalRequest.setInBoundLane(interAccessPoint);
-        assertEquals(signalRequest.getInBoundLane().getApproach(), 5);
+        assertEquals(signalRequest.getInBoundLane().getApproach().intValue(), 5);
 
         signalRequest.setOutBoundLane(interAccessPoint);
-        assertEquals(signalRequest.getOutBoundLane().getApproach(), 5);
+        assertEquals(signalRequest.getOutBoundLane().getApproach().intValue(), 5);
 	}
 }
