+++ jpo-ode-plugins/src/test/java/us/dot/its/jpo/ode/plugin/j2735/J2735SignalStatusTest.java
@@ -1,11 +1,11 @@
 package us.dot.its.jpo.ode.plugin.j2735;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.Assert.assertEquals;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import org.junit.jupiter.api.Test;
+import org.junit.Test;
 
 public class J2735SignalStatusTest {
     
@@ -14,12 +14,12 @@ public class J2735SignalStatusTest {
         J2735SignalStatus signalStatus = new J2735SignalStatus();
 
         signalStatus.setSequenceNumber(5);
-        assertEquals(signalStatus.getSequenceNumber(), 5);
+        assertEquals(signalStatus.getSequenceNumber().intValue(), 5);
 
         J2735IntersectionReferenceID interRefId = new J2735IntersectionReferenceID();
         interRefId.setId(5);
         signalStatus.setId(interRefId);
-        assertEquals(signalStatus.getId().getId(), 5);
+        assertEquals(signalStatus.getId().getId().intValue(), 5);
 
         J2735SignalStatusPackageList sigStatusList = new J2735SignalStatusPackageList();
         J2735SignalStatusPackage sigPackage = new J2735SignalStatusPackage();
